+++ sdk/policy/arm-policy/src/api/policyDefinitionVersions/operations.ts
@@ -4,13 +4,13 @@
 import type { PolicyContext as Client } from "../index.js";
 import type {
   PolicyDefinitionVersion,
-  _PolicyDefinitionVersionListResult,
+  PolicyDefinitionVersionListResult,
 } from "../../models/models.js";
 import {
   errorResponseDeserializer,
   policyDefinitionVersionSerializer,
   policyDefinitionVersionDeserializer,
-  _policyDefinitionVersionListResultDeserializer,
+  policyDefinitionVersionListResultDeserializer,
 } from "../../models/models.js";
 import type { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
 import { buildPagedAsyncIterator } from "../../static-helpers/pagingHelpers.js";
@@ -41,7 +41,7 @@ export function _listAllSend(
     "/subscriptions/{subscriptionId}/providers/Microsoft.Authorization/listPolicyDefinitionVersions{?api%2Dversion}",
     {
       subscriptionId: context.subscriptionId,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -55,22 +55,25 @@ export function _listAllSend(
 
 export async function _listAllDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicyDefinitionVersionListResult> {
+): Promise<PolicyDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policyDefinitionVersionListResultDeserializer(result.body);
+  return policyDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation lists all the policy definition versions for all policy definitions within a subscription. */
 export async function listAll(
   context: Client,
   options: PolicyDefinitionVersionsListAllOptionalParams = { requestOptions: {} },
-): Promise<_PolicyDefinitionVersionListResult> {
+): Promise<PolicyDefinitionVersionListResult> {
   const result = await _listAllSend(context, options);
   return _listAllDeserialize(result);
 }
@@ -84,7 +87,7 @@ export function _listAllAtManagementGroupSend(
     "/providers/Microsoft.Management/managementGroups/{managementGroupName}/providers/Microsoft.Authorization/listPolicyDefinitionVersions{?api%2Dversion}",
     {
       managementGroupName: managementGroupName,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -98,15 +101,18 @@ export function _listAllAtManagementGroupSend(
 
 export async function _listAllAtManagementGroupDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicyDefinitionVersionListResult> {
+): Promise<PolicyDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policyDefinitionVersionListResultDeserializer(result.body);
+  return policyDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation lists all the policy definition versions for all policy definitions at the management group scope. */
@@ -114,7 +120,7 @@ export async function listAllAtManagementGroup(
   context: Client,
   managementGroupName: string,
   options: PolicyDefinitionVersionsListAllAtManagementGroupOptionalParams = { requestOptions: {} },
-): Promise<_PolicyDefinitionVersionListResult> {
+): Promise<PolicyDefinitionVersionListResult> {
   const result = await _listAllAtManagementGroupSend(context, managementGroupName, options);
   return _listAllAtManagementGroupDeserialize(result);
 }
@@ -126,7 +132,7 @@ export function _listAllBuiltinsSend(
   const path = expandUrlTemplate(
     "/providers/Microsoft.Authorization/listPolicyDefinitionVersions{?api%2Dversion}",
     {
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -140,22 +146,25 @@ export function _listAllBuiltinsSend(
 
 export async function _listAllBuiltinsDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicyDefinitionVersionListResult> {
+): Promise<PolicyDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policyDefinitionVersionListResultDeserializer(result.body);
+  return policyDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation lists all the built-in policy definition versions for all built-in policy definitions. */
 export async function listAllBuiltins(
   context: Client,
   options: PolicyDefinitionVersionsListAllBuiltinsOptionalParams = { requestOptions: {} },
-): Promise<_PolicyDefinitionVersionListResult> {
+): Promise<PolicyDefinitionVersionListResult> {
   const result = await _listAllBuiltinsSend(context, options);
   return _listAllBuiltinsDeserialize(result);
 }
@@ -171,7 +180,7 @@ export function _listByManagementGroupSend(
     {
       managementGroupName: managementGroupName,
       policyDefinitionName: policyDefinitionName,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
       "%24top": options?.top,
     },
     {
@@ -186,15 +195,18 @@ export function _listByManagementGroupSend(
 
 export async function _listByManagementGroupDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicyDefinitionVersionListResult> {
+): Promise<PolicyDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policyDefinitionVersionListResultDeserializer(result.body);
+  return policyDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation retrieves a list of all the policy definition versions for the given policy definition in the given management group. */
@@ -209,7 +221,7 @@ export function listByManagementGroup(
     () => _listByManagementGroupSend(context, managementGroupName, policyDefinitionName, options),
     _listByManagementGroupDeserialize,
     ["200"],
-    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2025-03-01" },
+    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2026-07-01" },
   );
 }
 
@@ -226,7 +238,7 @@ export function _deleteAtManagementGroupSend(
       managementGroupName: managementGroupName,
       policyDefinitionName: policyDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -241,7 +253,10 @@ export async function _deleteAtManagementGroupDeserialize(
   const expectedStatuses = ["200", "204"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -282,7 +297,7 @@ export function _createOrUpdateAtManagementGroupSend(
       managementGroupName: managementGroupName,
       policyDefinitionName: policyDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -302,7 +317,10 @@ export async function _createOrUpdateAtManagementGroupDeserialize(
   const expectedStatuses = ["200", "201"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -344,7 +362,7 @@ export function _getAtManagementGroupSend(
       managementGroupName: managementGroupName,
       policyDefinitionName: policyDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -362,7 +380,10 @@ export async function _getAtManagementGroupDeserialize(
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -396,7 +417,7 @@ export function _listBuiltInSend(
     "/providers/Microsoft.Authorization/policyDefinitions/{policyDefinitionName}/versions{?api%2Dversion,%24top}",
     {
       policyDefinitionName: policyDefinitionName,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
       "%24top": options?.top,
     },
     {
@@ -411,15 +432,18 @@ export function _listBuiltInSend(
 
 export async function _listBuiltInDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicyDefinitionVersionListResult> {
+): Promise<PolicyDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policyDefinitionVersionListResultDeserializer(result.body);
+  return policyDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation retrieves a list of all the built-in policy definition versions for the given policy definition. */
@@ -433,7 +457,7 @@ export function listBuiltIn(
     () => _listBuiltInSend(context, policyDefinitionName, options),
     _listBuiltInDeserialize,
     ["200"],
-    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2025-03-01" },
+    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2026-07-01" },
   );
 }
 
@@ -448,7 +472,7 @@ export function _getBuiltInSend(
     {
       policyDefinitionName: policyDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -466,7 +490,10 @@ export async function _getBuiltInDeserialize(
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -499,7 +526,7 @@ export function _listSend(
     {
       subscriptionId: context.subscriptionId,
       policyDefinitionName: policyDefinitionName,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
       "%24top": options?.top,
     },
     {
@@ -514,15 +541,18 @@ export function _listSend(
 
 export async function _listDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicyDefinitionVersionListResult> {
+): Promise<PolicyDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policyDefinitionVersionListResultDeserializer(result.body);
+  return policyDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation retrieves a list of all the policy definition versions for the given policy definition. */
@@ -536,7 +566,7 @@ export function list(
     () => _listSend(context, policyDefinitionName, options),
     _listDeserialize,
     ["200"],
-    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2025-03-01" },
+    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2026-07-01" },
   );
 }
 
@@ -552,7 +582,7 @@ export function _$deleteSend(
       subscriptionId: context.subscriptionId,
       policyDefinitionName: policyDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -565,7 +595,10 @@ export async function _$deleteDeserialize(result: PathUncheckedResponse): Promis
   const expectedStatuses = ["200", "204"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -573,11 +606,6 @@ export async function _$deleteDeserialize(result: PathUncheckedResponse): Promis
 }
 
 /** This operation deletes the policy definition version in the given subscription with the given name. */
-/**
- *  @fixme delete is a reserved word that cannot be used as an operation name.
- *         Please add @clientName("clientName") or @clientName("<JS-Specific-Name>", "javascript")
- *         to the operation to override the generated name.
- */
 export async function $delete(
   context: Client,
   policyDefinitionName: string,
@@ -606,7 +634,7 @@ export function _createOrUpdateSend(
       subscriptionId: context.subscriptionId,
       policyDefinitionName: policyDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -626,7 +654,10 @@ export async function _createOrUpdateDeserialize(
   const expectedStatuses = ["200", "201"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -663,7 +694,7 @@ export function _getSend(
       subscriptionId: context.subscriptionId,
       policyDefinitionName: policyDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -681,7 +712,10 @@ export async function _getDeserialize(
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }