+++ sdk/policy/arm-policy/src/api/policySetDefinitionVersions/operations.ts
@@ -4,13 +4,13 @@
 import type { PolicyContext as Client } from "../index.js";
 import type {
   PolicySetDefinitionVersion,
-  _PolicySetDefinitionVersionListResult,
+  PolicySetDefinitionVersionListResult,
 } from "../../models/models.js";
 import {
   errorResponseDeserializer,
   policySetDefinitionVersionSerializer,
   policySetDefinitionVersionDeserializer,
-  _policySetDefinitionVersionListResultDeserializer,
+  policySetDefinitionVersionListResultDeserializer,
 } from "../../models/models.js";
 import type { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
 import { buildPagedAsyncIterator } from "../../static-helpers/pagingHelpers.js";
@@ -41,7 +41,7 @@ export function _listAllSend(
     "/subscriptions/{subscriptionId}/providers/Microsoft.Authorization/listPolicySetDefinitionVersions{?api%2Dversion}",
     {
       subscriptionId: context.subscriptionId,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -55,22 +55,25 @@ export function _listAllSend(
 
 export async function _listAllDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicySetDefinitionVersionListResult> {
+): Promise<PolicySetDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policySetDefinitionVersionListResultDeserializer(result.body);
+  return policySetDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation lists all the policy set definition versions for all policy set definitions within a subscription. */
 export async function listAll(
   context: Client,
   options: PolicySetDefinitionVersionsListAllOptionalParams = { requestOptions: {} },
-): Promise<_PolicySetDefinitionVersionListResult> {
+): Promise<PolicySetDefinitionVersionListResult> {
   const result = await _listAllSend(context, options);
   return _listAllDeserialize(result);
 }
@@ -86,7 +89,7 @@ export function _listAllAtManagementGroupSend(
     "/providers/Microsoft.Management/managementGroups/{managementGroupName}/providers/Microsoft.Authorization/listPolicySetDefinitionVersions{?api%2Dversion}",
     {
       managementGroupName: managementGroupName,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -100,15 +103,18 @@ export function _listAllAtManagementGroupSend(
 
 export async function _listAllAtManagementGroupDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicySetDefinitionVersionListResult> {
+): Promise<PolicySetDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policySetDefinitionVersionListResultDeserializer(result.body);
+  return policySetDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation lists all the policy set definition versions for all policy set definitions at the management group scope. */
@@ -118,7 +124,7 @@ export async function listAllAtManagementGroup(
   options: PolicySetDefinitionVersionsListAllAtManagementGroupOptionalParams = {
     requestOptions: {},
   },
-): Promise<_PolicySetDefinitionVersionListResult> {
+): Promise<PolicySetDefinitionVersionListResult> {
   const result = await _listAllAtManagementGroupSend(context, managementGroupName, options);
   return _listAllAtManagementGroupDeserialize(result);
 }
@@ -130,7 +136,7 @@ export function _listAllBuiltinsSend(
   const path = expandUrlTemplate(
     "/providers/Microsoft.Authorization/listPolicySetDefinitionVersions{?api%2Dversion}",
     {
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -144,22 +150,25 @@ export function _listAllBuiltinsSend(
 
 export async function _listAllBuiltinsDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicySetDefinitionVersionListResult> {
+): Promise<PolicySetDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policySetDefinitionVersionListResultDeserializer(result.body);
+  return policySetDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation lists all the built-in policy set definition versions for all built-in policy set definitions. */
 export async function listAllBuiltins(
   context: Client,
   options: PolicySetDefinitionVersionsListAllBuiltinsOptionalParams = { requestOptions: {} },
-): Promise<_PolicySetDefinitionVersionListResult> {
+): Promise<PolicySetDefinitionVersionListResult> {
   const result = await _listAllBuiltinsSend(context, options);
   return _listAllBuiltinsDeserialize(result);
 }
@@ -175,7 +184,7 @@ export function _listByManagementGroupSend(
     {
       managementGroupName: managementGroupName,
       policySetDefinitionName: policySetDefinitionName,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
       "%24expand": options?.expand,
       "%24top": options?.top,
     },
@@ -191,15 +200,18 @@ export function _listByManagementGroupSend(
 
 export async function _listByManagementGroupDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicySetDefinitionVersionListResult> {
+): Promise<PolicySetDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policySetDefinitionVersionListResultDeserializer(result.body);
+  return policySetDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation retrieves a list of all the policy set definition versions for the given policy set definition in a given management group. */
@@ -215,7 +227,7 @@ export function listByManagementGroup(
       _listByManagementGroupSend(context, managementGroupName, policySetDefinitionName, options),
     _listByManagementGroupDeserialize,
     ["200"],
-    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2025-03-01" },
+    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2026-07-01" },
   );
 }
 
@@ -234,7 +246,7 @@ export function _deleteAtManagementGroupSend(
       managementGroupName: managementGroupName,
       policySetDefinitionName: policySetDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -249,7 +261,10 @@ export async function _deleteAtManagementGroupDeserialize(
   const expectedStatuses = ["200", "204"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -292,7 +307,7 @@ export function _createOrUpdateAtManagementGroupSend(
       managementGroupName: managementGroupName,
       policySetDefinitionName: policySetDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -312,7 +327,10 @@ export async function _createOrUpdateAtManagementGroupDeserialize(
   const expectedStatuses = ["200", "201"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -354,7 +372,7 @@ export function _getAtManagementGroupSend(
       managementGroupName: managementGroupName,
       policySetDefinitionName: policySetDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
       "%24expand": options?.expand,
     },
     {
@@ -373,7 +391,10 @@ export async function _getAtManagementGroupDeserialize(
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -407,7 +428,7 @@ export function _listBuiltInSend(
     "/providers/Microsoft.Authorization/policySetDefinitions/{policySetDefinitionName}/versions{?api%2Dversion,%24expand,%24top}",
     {
       policySetDefinitionName: policySetDefinitionName,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
       "%24expand": options?.expand,
       "%24top": options?.top,
     },
@@ -423,15 +444,18 @@ export function _listBuiltInSend(
 
 export async function _listBuiltInDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicySetDefinitionVersionListResult> {
+): Promise<PolicySetDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policySetDefinitionVersionListResultDeserializer(result.body);
+  return policySetDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation retrieves a list of all the built-in policy set definition versions for the given built-in policy set definition. */
@@ -445,7 +469,7 @@ export function listBuiltIn(
     () => _listBuiltInSend(context, policySetDefinitionName, options),
     _listBuiltInDeserialize,
     ["200"],
-    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2025-03-01" },
+    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2026-07-01" },
   );
 }
 
@@ -460,7 +484,7 @@ export function _getBuiltInSend(
     {
       policySetDefinitionName: policySetDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
       "%24expand": options?.expand,
     },
     {
@@ -479,7 +503,10 @@ export async function _getBuiltInDeserialize(
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -512,7 +539,7 @@ export function _listSend(
     {
       subscriptionId: context.subscriptionId,
       policySetDefinitionName: policySetDefinitionName,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
       "%24expand": options?.expand,
       "%24top": options?.top,
     },
@@ -528,15 +555,18 @@ export function _listSend(
 
 export async function _listDeserialize(
   result: PathUncheckedResponse,
-): Promise<_PolicySetDefinitionVersionListResult> {
+): Promise<PolicySetDefinitionVersionListResult> {
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
-  return _policySetDefinitionVersionListResultDeserializer(result.body);
+  return policySetDefinitionVersionListResultDeserializer(result.body);
 }
 
 /** This operation retrieves a list of all the policy set definition versions for the given policy set definition. */
@@ -550,7 +580,7 @@ export function list(
     () => _listSend(context, policySetDefinitionName, options),
     _listDeserialize,
     ["200"],
-    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2025-03-01" },
+    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2026-07-01" },
   );
 }
 
@@ -566,7 +596,7 @@ export function _$deleteSend(
       subscriptionId: context.subscriptionId,
       policySetDefinitionName: policySetDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -579,7 +609,10 @@ export async function _$deleteDeserialize(result: PathUncheckedResponse): Promis
   const expectedStatuses = ["200", "204"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -587,11 +620,6 @@ export async function _$deleteDeserialize(result: PathUncheckedResponse): Promis
 }
 
 /** This operation deletes the policy set definition version in the given subscription with the given name and version. */
-/**
- *  @fixme delete is a reserved word that cannot be used as an operation name.
- *         Please add @clientName("clientName") or @clientName("<JS-Specific-Name>", "javascript")
- *         to the operation to override the generated name.
- */
 export async function $delete(
   context: Client,
   policySetDefinitionName: string,
@@ -620,7 +648,7 @@ export function _createOrUpdateSend(
       subscriptionId: context.subscriptionId,
       policySetDefinitionName: policySetDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -640,7 +668,10 @@ export async function _createOrUpdateDeserialize(
   const expectedStatuses = ["200", "201"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }
 
@@ -677,7 +708,7 @@ export function _getSend(
       subscriptionId: context.subscriptionId,
       policySetDefinitionName: policySetDefinitionName,
       policyDefinitionVersion: policyDefinitionVersion,
-      "api%2Dversion": context.apiVersion ?? "2025-03-01",
+      "api%2Dversion": context.apiVersion ?? "2026-07-01",
       "%24expand": options?.expand,
     },
     {
@@ -696,7 +727,10 @@ export async function _getDeserialize(
   const expectedStatuses = ["200"];
   if (!expectedStatuses.includes(result.status)) {
     const error = createRestError(result);
-    error.details = errorResponseDeserializer(result.body);
+    if (result.body) {
+      error.details = errorResponseDeserializer(result.body);
+    }
+
     throw error;
   }