+++ sdk/policy/arm-policy/src/classic/policyAssignments/index.ts
@@ -51,16 +51,11 @@ export interface PolicyAssignmentsOperations {
     options?: PolicyAssignmentsListForResourceGroupOptionalParams,
   ) => PagedAsyncIterableIterator<PolicyAssignment>;
   /** This operation deletes a policy assignment, given its name and the scope it was created in. The scope of a policy assignment is the part of its ID preceding '/providers/Microsoft.Authorization/policyAssignments/{policyAssignmentName}'. */
-  /**
-   *  @fixme delete is a reserved word that cannot be used as an operation name.
-   *         Please add @clientName("clientName") or @clientName("<JS-Specific-Name>", "javascript")
-   *         to the operation to override the generated name.
-   */
   delete: (
     scope: string,
     policyAssignmentName: string,
     options?: PolicyAssignmentsDeleteOptionalParams,
-  ) => Promise<PolicyAssignment>;
+  ) => Promise<PolicyAssignment | void>;
   /** This operation updates a policy assignment with the given scope and name. Policy assignments apply to all resources contained within their scope. For example, when you assign a policy at resource group scope, that policy applies to all resources in the group. */
   update: (
     scope: string,